+++ directives/extended-table.js
@@ -1087,7 +1087,8 @@ angular.module('keta.directives.ExtendedTable')
 '									<span class="glyphicon glyphicon-minus-sign"></span>' +
 '								</a>' +
 '							</th>' +
-'							<th data-ng-if="actionList.length">' +
+'							<th data-ng-if="actionList.length"' +
+'								class="{{columnClassCallback(headers, \'actions\', true)}}">' +
 '								{{headerLabelCallback(\'actions\')}}' +
 '							</th>' +
 '						</tr>' +
@@ -1102,7 +1103,8 @@ angular.module('keta.directives.ExtendedTable')
 '								class="{{columnClassCallback(row, column, false)}}">' +
 '								<span data-ng-bind-html="cellRenderer(row, column)"></span>' +
 '							</td>' +
-'							<td data-ng-if="row && actionList.length">' +
+'							<td data-ng-if="row && actionList.length"' +
+'								class="{{columnClassCallback(row, \'actions\', false)}}">' +
 '								<div class="btn-group" role="group">' +
 '									<span data-ng-repeat="item in actionList"' +
 '										data-ng-if="showActionListItem(item, row)">' +
@@ -1134,7 +1136,8 @@ angular.module('keta.directives.ExtendedTable')
 '								class="{{columnClassCallback(row, column, false)}}">' +
 '								<span data-ng-bind-html="cellRenderer(row, column)"></span>' +
 '							</td>' +
-'							<td data-ng-if="row && actionList.length">' +
+'							<td data-ng-if="row && actionList.length"' +
+'								class="{{columnClassCallback(row, \'actions\', false)}}">' +
 '								<div class="btn-group" role="group">' +
 '									<span data-ng-repeat="item in actionList"' +
 '										data-ng-if="showActionListItem(item, row)">' +